server/requirements.txt
Flask==3.1.0
gunicorn==23.0.0
numpy==2.2.5
scipy==1.15.2
skikit-image==0.25.2
pillow=11.1.0